feat: Add ChartReport class for generating wordclouds and bar charts
This commit adds a new class called ChartReport to the reports module. The ChartReport class provides methods for generating wordclouds and bar charts. It also includes a method for reading stopwords from a file. This class will be used to enhance the reporting capabilities of the application.

The ChartReport class has the following methods:
- wordcloud: Generates a wordcloud from a given text.
- most_common_tags: Generates a bar chart of the most common tags.

The stopwords.txt file is also added to the reports module, which contains a list of common stopwords used for text analysis.

This commit addresses the need for visualizing data in the application and improves the reporting functionality.