Refactor TagGroup class and add TagsCounter class
This commit refactors the TagGroup class by removing it from the codebase. The TagGroup class represented a group of tags and had methods to add and remove tags. However, it was not being used and was causing unnecessary complexity. Therefore, it has been deleted.

In addition, this commit introduces a new TagsCounter class. The TagsCounter class is responsible for counting the occurrences of tags. It has methods to increment, decrement, and retrieve the count of a tag. This class will be used to manage the tag counts in the application.

These changes improve the code structure and provide a more efficient way to handle tag counts.